Move intermediate storage out of data collection, introduce new HDF store adapter
