+++ tests/dwd/test_index.py
@@ -5,16 +5,6 @@ from wetterdienst.dwd.metadata.period_type import PeriodType
 from wetterdienst import TimeResolution
 from wetterdienst.util.network import list_remote_files
 from wetterdienst.dwd.index import build_path_to_parameter
-from wetterdienst.dwd.observations.store import build_local_filepath_for_station_data
-
-
-def test_build_local_filepath_for_station_data():
-    local_filepath = build_local_filepath_for_station_data("dwd_data")
-
-    assert (
-        "/".join(local_filepath.as_posix().split("/")[-3:])
-        == "dwd_data/station_data/dwd_station_data.h5"
-    )
 
 
 def test_build_index_path():
