+++ tests/dwd/test_util.py
@@ -7,7 +7,7 @@ from wetterdienst.dwd.util import (
     check_parameters,
     coerce_field_types,
     parse_enumeration_from_template,
-    create_humanized_column_names_mapping,
+    build_parameter_identifier,
 )
 from wetterdienst.dwd.metadata.period_type import PeriodType
 from wetterdienst.dwd.metadata.time_resolution import TimeResolution
@@ -82,27 +82,9 @@ def test_coerce_field_types_with_nans():
     assert_frame_equal(df, expected_df)
 
 
-def test_create_humanized_column_names_mapping():
-    """ Test for function to create a mapping to humanized column names """
-    hcnm = create_humanized_column_names_mapping(
-        TimeResolution.DAILY, Parameter.CLIMATE_SUMMARY
+def test_build_parameter_identifier():
+    parameter_identifier = build_parameter_identifier(
+        Parameter.CLIMATE_SUMMARY, TimeResolution.DAILY, PeriodType.HISTORICAL, 1
     )
 
-    assert hcnm == {
-        "QN_3": "QUALITY_WIND",
-        "FX": "WIND_GUST_MAX",
-        "FM": "WIND_VELOCITY",
-        "QN_4": "QUALITY_GENERAL",
-        "RSK": "PRECIPITATION_HEIGHT",
-        "RSKF": "PRECIPITATION_FORM",
-        "SDK": "SUNSHINE_DURATION",
-        "SHK_TAG": "SNOW_DEPTH",
-        "NM": "CLOUD_COVERAGE_TOTAL",
-        "VPM": "PRESSURE_VAPOR",
-        "PM": "PRESSURE_AIR",
-        "TMK": "TEMPERATURE_AIR_200",
-        "UPM": "HUMIDITY",
-        "TXK": "TEMPERATURE_AIR_MAX_200",
-        "TNK": "TEMPERATURE_AIR_MIN_200",
-        "TGK": "TEMPERATURE_AIR_MIN_005",
-    }
+    assert parameter_identifier == "kl/daily/historical/station_id_1"
