+++ wetterdienst/dwd/observations/access.py
@@ -3,7 +3,7 @@ import logging
 from concurrent.futures.thread import ThreadPoolExecutor
 from io import BytesIO
 from pathlib import Path
-from typing import List, Union, Optional, Tuple
+from typing import List, Union, Tuple
 from zipfile import ZipFile, BadZipFile
 
 import pandas as pd
@@ -17,14 +17,12 @@ from wetterdienst.util.cache import payload_cache_five_minutes
 from wetterdienst.dwd.util import (
     check_parameters,
     parse_enumeration_from_template,
-    create_humanized_column_names_mapping,
     coerce_field_types,
+    build_parameter_identifier,
 )
-from wetterdienst.dwd.metadata.column_names import DWDMetaColumns
 from wetterdienst.dwd.metadata.parameter import Parameter
 from wetterdienst.dwd.metadata.period_type import PeriodType
 from wetterdienst import TimeResolution
-from wetterdienst.dwd.metadata.constants import DWD_FOLDER_MAIN
 from wetterdienst.exceptions import (
     InvalidParameterCombination,
     FailedDownload,
@@ -34,62 +32,29 @@ from wetterdienst.dwd.observations.parser import (
     parse_climate_observations_data,
 )
 from wetterdienst.dwd.network import download_file_from_dwd
-from wetterdienst.dwd.observations.store import (
-    store_climate_observations,
-    restore_climate_observations,
-    _build_local_store_key,
-)
 
 log = logging.getLogger(__name__)
-
-
-POSSIBLE_ID_VARS = (
-    DWDMetaColumns.STATION_ID.value,
-    DWDMetaColumns.DATE.value,
-    DWDMetaColumns.FROM_DATE.value,
-    DWDMetaColumns.TO_DATE.value,
-)
-
-POSSIBLE_DATE_VARS = (
-    DWDMetaColumns.DATE.value,
-    DWDMetaColumns.FROM_DATE.value,
-    DWDMetaColumns.TO_DATE.value,
-)
 
 PRODUCT_FILE_IDENTIFIER = "produkt"
 
 
 def collect_climate_observations_data(
-    station_ids: List[int],
+    station_id: int,
     parameter: Union[Parameter, str],
     time_resolution: Union[TimeResolution, str],
     period_type: Union[PeriodType, str],
-    folder: Union[str, Path] = DWD_FOLDER_MAIN,
-    prefer_local: bool = False,
-    write_file: bool = False,
-    tidy_data: bool = True,
-    humanize_column_names: bool = False,
-    run_download_only: bool = False,
-) -> Optional[pd.DataFrame]:
+) -> pd.DataFrame:
     """
     Function that organizes the complete pipeline of data collection, either
-    from the internet or from a local file. It therefor goes through every given
+    from the internet or from a local file. It therefore goes through every given
     station id and, given by the parameters, either tries to get data from local
     store and/or if fails tries to get data from the internet. Finally if wanted
     it will try to store the data in a hdf file.
 
-    :param station_ids:             station ids that are trying to be loaded
+    :param station_id:              station id that is being loaded
     :param parameter:               Parameter as enumeration
     :param time_resolution:         Time resolution as enumeration
     :param period_type:             Period type as enumeration
-    :param folder:                  Folder for local file interaction
-    :param prefer_local:            Local data should be preferred
-    :param write_file:              Write data to local storage
-    :param tidy_data:               Tidy up data so that there's only one set of values
-                                    for a datetime in a row, e.g. station_id, parameter,
-                                    element, datetime, value, quality.
-    :param humanize_column_names:   Yield column names for human consumption
-    :param run_download_only:       Run only the download and storing process
 
     :return:                        All the data given by the station ids.
     """
@@ -99,143 +64,30 @@ def collect_climate_observations_data(
 
     if not check_parameters(parameter, time_resolution, period_type):
         raise InvalidParameterCombination(
-            f"The combination of {parameter.value}, {time_resolution.value}, "
-            f"{period_type.value} is invalid."
+            f"Invalid combination: {parameter.value} / {time_resolution.value} / "
+            f"{period_type.value}"
         )
 
-    # List for collected pandas DataFrames per each station id
-    data = []
-    for station_id in set(station_ids):
+    remote_files = create_file_list_for_climate_observations(
+        station_id, parameter, time_resolution, period_type
+    )
 
-        # Just for logging.
-        request_string = _build_local_store_key(
-            station_id, parameter, time_resolution, period_type
+    if len(remote_files) == 0:
+        parameter_identifier = build_parameter_identifier(
+            parameter, time_resolution, period_type, station_id
         )
-
-        if prefer_local:
-            # Try restoring data
-            station_data = restore_climate_observations(
-                station_id, parameter, time_resolution, period_type, folder
-            )
-
-            # When successful append data and continue with next iteration
-            if not station_data.empty:
-                log.info(f"Data for {request_string} restored from local.")
-
-                data.append(station_data)
-
-                continue
-
-        log.info(f"Acquiring observations data for {request_string}")
-
-        remote_files = create_file_list_for_climate_observations(
-            [station_id], parameter, time_resolution, period_type
-        )
-
-        if len(remote_files) == 0:
-            log.info(f"No files found for {request_string}. Station will be skipped.")
-            continue
-
-        filenames_and_files = download_climate_observations_data_parallel(remote_files)
-
-        station_data = parse_climate_observations_data(
-            filenames_and_files, parameter, time_resolution
-        )
-
-        if write_file:
-            store_climate_observations(
-                station_data,
-                station_id,
-                parameter,
-                time_resolution,
-                period_type,
-                folder,
-            )
-
-        data.append(station_data)
-
-    if run_download_only:
-        return None
-
-    try:
-        data = pd.concat(data)
-    except ValueError:
+        log.info(f"No files found for {parameter_identifier}. Station will be skipped.")
         return pd.DataFrame()
 
-    data = coerce_field_types(data, time_resolution)
+    filenames_and_files = download_climate_observations_data_parallel(remote_files)
 
-    if tidy_data:
-        data = _tidy_up_data(data, parameter)
-
-    # Assign meaningful column names (humanized).
-    if humanize_column_names:
-        hcnm = create_humanized_column_names_mapping(time_resolution, parameter)
-        if tidy_data:
-            data[DWDMetaColumns.ELEMENT.value] = data[
-                DWDMetaColumns.ELEMENT.value
-            ].apply(lambda x: hcnm[x])
-        else:
-            data = data.rename(columns=hcnm)
-
-    return data
-
-
-def _tidy_up_data(df: pd.DataFrame, parameter: Parameter) -> pd.DataFrame:
-    """
-    Function to create a tidy DataFrame by reshaping it, putting quality in a
-    separate column and setting an extra column with the parameter.
-
-    :param df:          DataFrame to be tidied
-    :param parameter:   the parameter that is written in a column to identify a set of
-                        different parameters amongst each other
-
-    :return:            The tidied DataFrame
-    """
-    id_vars = []
-    date_vars = []
-
-    # Add id columns based on metadata columns
-    for column in POSSIBLE_ID_VARS:
-        if column in df:
-            id_vars.append(column)
-            if column in POSSIBLE_DATE_VARS:
-                date_vars.append(column)
-
-    # Extract quality
-    # Set empty quality for first columns until first QN column
-    quality = pd.Series(dtype=int)
-    column_quality = pd.Series(dtype=int)
-
-    for column in df:
-        # If is quality column, overwrite current "column quality"
-        if column.startswith("QN"):
-            column_quality = df.pop(column)
-        else:
-            quality = quality.append(column_quality)
-
-    df_tidy = df.melt(
-        id_vars=id_vars,
-        var_name=DWDMetaColumns.ELEMENT.value,
-        value_name=DWDMetaColumns.VALUE.value,
+    obs_df = parse_climate_observations_data(
+        filenames_and_files, parameter, time_resolution
     )
 
-    df_tidy[DWDMetaColumns.PARAMETER.value] = parameter.name
+    obs_df = coerce_field_types(obs_df, time_resolution)
 
-    df_tidy[DWDMetaColumns.QUALITY.value] = quality.values
-
-    # Reorder properly
-    df_tidy = df_tidy.reindex(
-        columns=[
-            DWDMetaColumns.STATION_ID.value,
-            DWDMetaColumns.PARAMETER.value,
-            DWDMetaColumns.ELEMENT.value,
-            *date_vars,
-            DWDMetaColumns.VALUE.value,
-            DWDMetaColumns.QUALITY.value,
-        ]
-    )
-
-    return df_tidy
+    return obs_df
 
 
 def download_climate_observations_data_parallel(
@@ -279,7 +131,7 @@ def __download_climate_observations_data(remote_file: str) -> bytes:
         zip_file = download_file_from_dwd(remote_file)
     except InvalidURL as e:
         raise InvalidURL(
-            f"Error: the station data {remote_file} couldn't be reached."
+            f"Error: the station data {remote_file} could not be reached."
         ) from e
     except Exception:
         raise FailedDownload(f"Download failed for {remote_file}")
